+++ platform/native_libetc.c
@@ -12,7 +12,10 @@ global_variable int s_videoMode = -1;
 
 int VSyncCallback(void (*func)(void))
 {
-	int old = (int)vsync_callback;
+	// NOTE: return value is a retail 32-bit function-pointer-as-int; no
+	// caller uses it, so round-trip through uintptr_t instead of casting the
+	// pointer directly to int, which truncates on 64-bit targets.
+	int old = (int)(uintptr_t)vsync_callback;
 
 	vsync_callback = func;
 	return old;
@@ -25,7 +28,10 @@ int StopCallback(void)
 
 int ResetCallback(void)
 {
-	int old = (int)vsync_callback;
+	// NOTE: return value is a retail 32-bit function-pointer-as-int; no
+	// caller uses it, so round-trip through uintptr_t instead of casting the
+	// pointer directly to int, which truncates on 64-bit targets.
+	int old = (int)(uintptr_t)vsync_callback;
 
 	vsync_callback = NULL;
 	return old;
